requirements.txt
bs4
glove_python
nltk
numpy
wget
